refactor: user info pages
