+++ src/components/userinfo/index.ts
@@ -1,0 +1,3 @@
+export { default as UserInfoHeader } from './Header';
+export { default as UserInfoCardContainer } from './UserInfoCardContainer';
+export { default as UserInfoContainer } from './UserInfoContainer';
