+++ src/pages/userInfo/UserReviewPage.tsx
@@ -1,24 +1,15 @@
-import MyPageHeader from '@/components/userinfo/Header';
-import UserInfoContainer from '@/components/userinfo/UserInfoContainer';
-import { InfoSection, PlaceImage, PlaceInfo, UserInfoCard } from '@/components/userinfo/card';
+import { UserInfoCardContainer, UserInfoContainer, UserInfoHeader } from '@/components/userinfo';
+import { UserInfoCard } from '@/components/userinfo/card';
 
-const UserReviewPage = () => {
+const UserFavouritePage = () => {
   return (
-    <div className="flex h-full flex-col space-y-8 px-[4rem] py-8 md:py-4 lg:px-[12rem]">
-      <MyPageHeader text="리뷰"></MyPageHeader>
-      <UserInfoContainer>
-        <UserInfoCard>
-          <PlaceImage
-            src="https://www.travelonline.com/thailand/attractions/white-temple/thailand-generic-white-temple-hd-56217-banner.jpg"
-            alt="White Temple"
-          />
-          <InfoSection>
-            <PlaceInfo />
-          </InfoSection>
-        </UserInfoCard>
-      </UserInfoContainer>
-    </div>
+    <UserInfoContainer>
+      <UserInfoHeader text="리뷰"></UserInfoHeader>
+      <UserInfoCardContainer>
+        <UserInfoCard />
+      </UserInfoCardContainer>
+    </UserInfoContainer>
   );
 };
 
-export default UserReviewPage;
+export default UserFavouritePage;
